Moved types into the src folder
